希尔排序 插入排序 冒泡排序

+++ 插入排序/插入排序/test.cc
@@ -3,14 +3,14 @@
 
 using namespace std;
 
-void InsertionSort(int *arr, size_t len){
+void InsertionSort(int *arr, int len){
 	if (arr == NULL)
 		return;
 	if (len <= 1)
 		return;
 
-	for (size_t i = 0; i < len - 1; i++){
-		for (size_t j = i; j >= 0; j--){
+	for (int i = 0; i < len - 1; i++){
+		for (int j = i; j >= 0; j--){
 			if (arr[j] > arr[j + 1])
 				swap(arr[j], arr[j + 1]);
 			//来到else说明到这个位置已经有序了，所以break
@@ -23,7 +23,7 @@ void InsertionSort(int *arr, size_t len){
 
 int main(){
 	int arr[] = { 10, 9, 8, 7, 6, 5, 4, 3, 2, 1 };
-	size_t len = 10;
+	int len = 10;
 	InsertionSort(arr, len);
 	for (size_t i = 0; i < 10; i++){
 		cout << arr[i];
